chore(repo): revamp package publishing
